# tuma-platform/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuma API",
    "description": "Backend API for the Tuma commerce + delivery platform",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/merchant/stores": {
      "get": {
        "tags": [
          "merchant"
        ],
        "operationId": "list_own_stores",
        "responses": {
          "200": {
            "description": "The stores this operator can reach, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "No merchant membership"
          }
        }
      },
      "post": {
        "tags": [
          "merchant"
        ],
        "operationId": "create_own_store",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Store created (starts closed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Account belongs to several businesses"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "No merchant membership, or not the business owner"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/merchant/stores/{id}": {
      "get": {
        "tags": [
          "merchant"
        ],
        "operationId": "get_own_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One of this operator's stores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "No merchant membership"
          },
          "404": {
            "description": "Not one of this operator's stores"
          }
        }
      },
      "delete": {
        "tags": [
          "merchant"
        ],
        "summary": "Hard-delete one of the operator's stores. Its store_products follow via\nON DELETE CASCADE — permanent, and the platform confirms before calling.\nOwners only: a store-scoped manager cannot delete their store.",
        "operationId": "delete_own_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Store deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "No merchant membership, or not the business owner"
          },
          "404": {
            "description": "Not one of this operator's stores"
          }
        }
      },
      "patch": {
        "tags": [
          "merchant"
        ],
        "operationId": "update_own_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated store",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "No merchant membership"
          },
          "404": {
            "description": "Not one of this operator's stores"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/v1/stores": {
      "get": {
        "tags": [
          "stores"
        ],
        "operationId": "list_stores",
        "responses": {
          "200": {
            "description": "Open stores, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not a customer"
          }
        }
      }
    },
    "/v1/stores/{id}": {
      "get": {
        "tags": [
          "stores"
        ],
        "operationId": "get_store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The store and its menu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreWithProductsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not a customer"
          },
          "404": {
            "description": "No open store with that id"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateStoreInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "address_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ],
            "description": "What the store sells, in one word or two (e.g. \"Grill\", \"Bakery\")."
          },
          "delivery_fee": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Integer RWF. Defaults to 0 (free delivery)."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "lat": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "lng": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "MenuItemResponse": {
        "type": "object",
        "description": "A placeholder until the S3 catalog slice lands: the store detail's menu\nis empty because store_products have no endpoints yet. The mobile store\nscreen renders the honest empty state.",
        "required": [
          "id",
          "name",
          "price",
          "is_available"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_available": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "StoreResponse": {
        "type": "object",
        "description": "A store as the API returns it. Integer RWF for money, ISO-8601 times.",
        "required": [
          "id",
          "merchant_id",
          "name",
          "delivery_fee",
          "is_open",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "address_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "delivery_fee": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_open": {
            "type": "boolean"
          },
          "lat": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "lng": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "merchant_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StoreWithProductsResponse": {
        "type": "object",
        "required": [
          "store",
          "products"
        ],
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuItemResponse"
            }
          },
          "store": {
            "$ref": "#/components/schemas/StoreResponse"
          }
        }
      },
...
      "UpdateStoreInput": {
        "type": "object",
        "properties": {
          "address_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "delivery_fee": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_open": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "lat": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "lng": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}